feat: retry failed PublicExportJobs
Before we "silently" failed exports if e.g. Virtuoso's checkpointing
caused our queries to fail. These failed jobs hung around in our
database but we never did anything with them. We suspect that some of
the users' complaints about publications not reaching Valvas was due to
this. Our first step towards fixing that is to ensure that we retry
failed jobs up to 5 times.

Further steps will be taken to alleviate the process, by making the
status of the export jobs more explicit so that users can track it.