feat: handle failed jobs in the same flow as scheduled jobs
The previous handling of failed jobs occurred in parallel to the normal
flow of scheduled jobs. Although this should in theory not cause any
errors, part of the specified behaviour of this service is that export
jobs are executed one at a time in a FIFO order. For that reason, we
move the retry mechanism into the regular flow, so that if a job fails
the regular flow will pick it up again first (since its creation date is
before the next scheduled job) and retry up to 5 times.