chore: tweak the query to fetch unpublished activities
Before we would get all activities inside a window of time and then
iterate over them and verify that no job existed for them (so as to not
double schedule them). If we instead immediately verify that they have
no job in the initial query, we save a bunch of duplicate work down the
line.

I suspect the initial approach only existed because of the separate
kaleidos and themis-export databases that existed (which is no longer
the case, so you can in one swoop query the publication activities and
the export jobs).